// src/app/core/models/teacher-dashboard-models/skillsActivation.model.ts
export class SkillActivation {
    courseSectionIdList: number[] = [];
    domainId: number = 0;
    semesterId: number = 0;
    currentCourseSectionId: number = 0;
    lessonId: number = 0;
    curriculumLearningOutcomeId: number = 0;
    learningOutcomeId: number = 0;
    activationStatus: boolean;
  }